Amend to change into a unregistered subdirectory.
With this commit it's now possible to register a `parent` directory
and change into a unregistered subdirectory in one step.

e.g. In a `project` directory there are several subdirectories for
issues you are currently working on.

```
$ goto project issue*123
```

It will first change into the directory registered with the alias
`project` and from there to the directory matching `issue*123`.

This avoid the need to register every subdirectory of project. Especially
if the directory structure below `project` changes more frequently.